% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Brute force search}
\usage{
brute_force_knapsack(x, W)
}
\arguments{
\item{x}{a data.frame with two variables v and w}

\item{W}{the knapsack size, a positive number}
}
\value{
the maximum knapsack value and which elements (rows in the data.frame)
}
\description{
going through all possible alternatives and return the maximum value found
}
