% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack_parallel.R
\name{brute_force_knapsack_parallel}
\alias{brute_force_knapsack_parallel}
\title{Brute force search}
\usage{
brute_force_knapsack_parallel(x, W, parallel = FALSE)
}
\arguments{
\item{x}{a data.frame with two variables v and w}

\item{W}{the knapsack size, a positive number}

\item{parallel}{FALSE for non-parallel, TRUE for parallel,default F}
}
\value{
the maximum knapsack value and which elements (rows in the data.frame)
}
\description{
going through all possible alternatives and return the maximum value found, with a choice of parallel.
}
