% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{greddy heuristic}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{a data.frame with two variables v and w}

\item{W}{the knapsack size, a positive number}
}
\value{
the maximum knapsack value and which elements (rows in the data.frame)
}
\description{
greddy heuristic
}
\examples{
greedy_knapsack(x = knapsack_objects[1:8,], W = 3500)
}
