% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_brute_force.R
\name{knapsack_brute_force}
\alias{knapsack_brute_force}
\title{Knapsack Brute Force}
\usage{
knapsack_brute_force(x, W, parallel = FALSE)
}
\arguments{
\item{x}{Data frame containing values \code{v} and weights \code{w} of each item.}

\item{W}{Maximum capacity of the knapsack.}

\item{parallel}{(Optional) If the main loop should be run in parallel}
}
\value{
List containing the maximum value and the indices of the items.
}
\description{
A brute force (or exhaustive search) solution to the knapsack problem.
}
