% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{dynamic knapsack}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{a data.frame with two variables v and w}

\item{W}{the knapsack size, a positive number}
}
\value{
the maximum knapsack value and which elements (rows in the data.frame)
}
\description{
dynamic knapsack
}
