% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_greedy.R
\name{knapsack_greedy}
\alias{knapsack_greedy}
\title{Knapsack Greedy}
\usage{
knapsack_greedy(x, W)
}
\arguments{
\item{x}{Data frame containing values \code{v} and weights \code{w} of each item.}

\item{W}{Maximum capacity of the knapsack.}
}
\value{
List containing the maximum value and the indices of the items.
}
\description{
A greedy approximation to the knapsack problem where the aim is to pack the knapsack with the highest value to weight ratio items until the capacity is filled. It runs in \code{O(n log(n) + n)} due to the sorting algorithm (where \code{n log(n)} is the sorting, and \code{n} is the packing of items into the knapsack), but will not necessarily return the maximum value. The maximum value returned is greater than that of the test suite, because it doesn't pack the bag by weight, rather it packs it by the highest value to weight ratio. This, however, does not interfere with the complexity of the solution, i.e. the packing of items is done in O(n) time as can be seen in the function.
}
