% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knappack_object.R
\docType{data}
\name{knapsack_objects}
\alias{knapsack_objects}
\title{data for knappack}
\format{
a data.frame with 2000 obs of 2 variables

\describe{
  \item{w}{weight of the item}
  \item{v}{value of the item}
  }
}
\source{
https://www.ida.liu.se/~732A94/info/courseinfo.en.shtml lab6 2022
}
\usage{
knapsack_objects
}
\description{
The data we will use is generated with {set.seed(42, kind = "Mersenne-Twister", normal.kind = "Inversion")}. To create larger datasets, just set n to a larger number.
}
\keyword{datasets}
