\name{lab6package-package}
\alias{lab6package-package}
\alias{lab6package}
\docType{package}
\title{
\packageTitle{lab6package}
}
\description{
\packageDescription{lab6package}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{lab6package}
\packageIndices{lab6package}

}
\author{
\packageAuthor{lab6package}

Maintainer: \packageMaintainer{lab6package}
}

\keyword{ package }
