% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab6package-package.R
\docType{package}
\name{lab6package}
\alias{lab6package}
\alias{lab6package-package}
\title{Knapsack problem}
\description{
This package contains three functions, each trying to solve the knapsack problem in a different (and increasingly efficient) way.
These functions are \code{\link{knapsack_brute_force}}, \code{\link{knapsack_dynamic}} and \code{\link{knapsack_greedy}}.
}
\author{
\strong{Maintainer}: Amer Memic \email{ameme698@student.liu.se}

Authors:
\itemize{
  \item Olof Josefsson \email{olojo524@student.liu.se}
  \item Shahreen Saleem \email{shahreen.saleem111@gmail.com}
}

}
