% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{brute_force_knapsack}
\usage{
brute_force_knapsack(x, W, parallel)
}
\arguments{
\item{x}{Data frame.}

\item{W}{Value.}

\item{parallel}{Value.}
}
\value{
List
}
\description{
This is the brute force implementation of knapsack.
}
