% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{greedy_knapsack}
\usage{
greedy_knapsack  (x, W)
}
\arguments{
\item{x}{Data frame.}

\item{W}{Value.}
}
\value{
List
}
\description{
This is the greedy implementation of knapsack.
}
