% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Knapsack Dynamic Programming}
\usage{
knapsack_dynamic(x, W, fast = FALSE)
}
\arguments{
\item{x}{Data frame containing values \code{v} and weights \code{w} of each item.}

\item{W}{Maximum capacity of the knapsack.}

\item{fast}{(Optional) If the main loop should be run in C++}
}
\value{
List containing the maximum value and the indices of the items.
}
\description{
A dynamic programming approach to solve the knapsack problem.
}
