% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{knapsack_dynamic}
\usage{
knapsack_dynamic (x, W)
}
\arguments{
\item{x}{Data frame.}

\item{W}{Value.}
}
\value{
List
}
\description{
This is the dynamic implementation of knapsack.
}
